backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
beautifulsoup4
httpx
pydantic
python-dotenv
openai
requests
